feat:moved math-game game starting mechanism to service layer, removed outdated wait.time function when game end and extract the trigger points as CLI flag: play
