+++ math-game/internal/cobra/cobra.go
@@ -2,11 +2,12 @@ package cobra
 
 import (
 	"github.com/Anthya1104/math-game-cli/internal/config"
+	"github.com/Anthya1104/math-game-cli/internal/service"
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 )
 
-var maxRoundsFlag int
+var maxRounds int
 
 var rootCmd = &cobra.Command{
 	Use:   "app",
@@ -24,10 +25,20 @@ var versionCmd = &cobra.Command{
 	},
 }
 
+var playCmd = &cobra.Command{
+	Use:   "play",
+	Short: "Run math game play with input rounds",
+	Run: func(cmd *cobra.Command, args []string) {
+		service.StartGamePlay(maxRounds)
+	},
+}
+
 func InitCLI() *cobra.Command {
 
-	rootCmd.PersistentFlags().IntVarP(&maxRoundsFlag, "rounds", "r", 1, "Max game play round") // Maximum number of rounds for the game
+	playCmd.PersistentFlags().IntVarP(&maxRounds, "rounds", "r", 1, "Max game play round") // Maximum number of rounds for the game
+
 	rootCmd.AddCommand(versionCmd)
+	rootCmd.AddCommand(playCmd)
 
 	return rootCmd
 }
@@ -36,9 +47,4 @@ func ExecuteCmd() error {
 
 	return InitCLI().Execute()
 
-}
-
-// GetMaxRoundsFlag returns the value of the --rounds CLI flag.
-func GetMaxRoundsFlag() int {
-	return maxRoundsFlag
 }
